Revert "Estructura base de componente de estadisticas"
