--- frontend/src/assets/css/estadisticas.css
@@ -1,58 +1,0 @@
-.header-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-  max-width: 1100px;
-  margin-left: auto;
-  margin-right: auto;
-  padding-top: 24px;
-}
-
-.estadisticas-content {
-  background-color: #ffffff;
-  border-radius: 8px;
-  padding: 20px;
-  margin-top: 20px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.estadisticas-main-container {
-  width: 100%;
-  min-height: 100vh;
-  background: #fafbfc;
-}
-
-.estadisticas-content-centered {
-  max-width: 1100px;
-  margin: 0 auto;
-  padding: 32px 16px 64px 16px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.estadisticas-cards-grid {
-  width: 100%;
-  max-width: 900px;
-  margin: 0 auto 32px auto;
-}
-
-.estadisticas-chart-block {
-  width: 100%;
-  max-width: 900px;
-  margin: 0 auto 32px auto;
-  background: none;
-}
-
-@media (max-width: 1000px) {
-  .estadisticas-content-centered, .estadisticas-chart-block, .estadisticas-cards-grid {
-    max-width: 98vw;
-  }
-}
-
-@media (max-width: 600px) {
-  .estadisticas-content-centered, .estadisticas-chart-block, .estadisticas-cards-grid {
-    padding: 0 4px;
-  }
-}
