+++ frontend/src/pages/Dashboard.tsx
@@ -5,7 +5,6 @@ import EditRubric from "@/components/rubrics/EditRubric";
 import ConsultRubric from "@/components/rubrics/ConsultRubric";
 import RubricDetail from "@/components/rubrics/RubricDetail";
 import Evaluaciones from "@/views/Evaluaciones/Evaluation";
-import Estadisticas from "@/views/Estadisticas/Estadisticas";
 
 export default function Dashboard() {
     return (
@@ -17,7 +16,6 @@ export default function Dashboard() {
                 <Route path="/rubricas/editar/:id" element={<EditRubric />} />
                 <Route path="/rubricas/detalle/:id" element={<RubricDetail />} />
                 <Route path="/evaluaciones" element={<Evaluaciones />} />
-                <Route path="/estadisticas" element={<Estadisticas />} />
 
             </Routes>
         </DashboardLayout>
