--- frontend/src/views/Estadisticas/components/EstadisticasCards.css
@@ -1,64 +1,0 @@
-.estadisticas-cards-grid {
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(2, 1fr);
-  gap: 24px 32px;
-  margin-bottom: 40px;
-  justify-items: center;
-}
-
-.estadistica-card {
-  background: #f8f9fa;
-  border-radius: 12px;
-  box-shadow: 0 2px 8px rgba(44, 62, 80, 0.07);
-  padding: 24px 0 16px 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-height: 90px;
-  border: 2px solid #e3e6f0;
-  width: 95%;
-  max-width: 260px;
-}
-
-.estadistica-label {
-  font-size: 1.08rem;
-  color: #2e2ebe;
-  font-weight: 600;
-  margin-bottom: 8px;
-  text-align: center;
-}
-
-.estadistica-valor {
-  font-size: 2.2rem;
-  font-weight: 700;
-  color: #0d47a1;
-  text-align: center;
-}
-
-.card-max {
-  border-color: #b6e7c9;
-}
-
-.card-min {
-  border-color: #f7caca;
-}
-
-@media (max-width: 900px) {
-  .estadisticas-cards-grid {
-    grid-template-columns: 1fr 1fr;
-    grid-template-rows: repeat(3, 1fr);
-  }
-}
-
-@media (max-width: 600px) {
-  .estadisticas-cards-grid {
-    grid-template-columns: 1fr;
-    grid-template-rows: repeat(6, 1fr);
-    gap: 18px 0;
-  }
-  .estadistica-card {
-    max-width: 98vw;
-  }
-}
