--- frontend/src/views/Estadisticas/components/EstadisticasFilters.css
@@ -1,25 +1,0 @@
-.estadisticas-filters {
-  background-color: #ffffff;
-  border-radius: 8px;
-  padding: 20px;
-  margin-bottom: 20px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.filters-container {
-  display: flex;
-  gap: 20px;
-}
-
-.filter-block {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-@media (max-width: 768px) {
-  .filters-container {
-    flex-direction: column;
-  }
-}
