--- frontend/src/views/Estadisticas/types/index.ts
@@ -1,18 +1,0 @@
-export interface Rubrica {
-  idRubrica: number;
-  nombreRubrica: string;
-  materia: string;
-  objetivoEstudio: string;
-  courseId: number;
-}
-
-export interface Materia {
-  id: number;
-  name: string;
-}
-
-export interface EstadisticasFilters {
-  materia: string;
-  rubrica: string;
-  periodo: string;
-}
